pyproject.toml
[tool.poetry]
name = "coeusfactory"
version = "0.1.0"
description = ""
authors = ["amendoza <amendoza@stratpoint.com>"]

[tool.poetry.dependencies]
python = "^3.6"
rainbow-bridge-logger = "^0.6.0"
pypubsub = "^4.0.3"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
flake8 = "^3.7.9"
boto3 = "^1.11.15"
pymongo = "^3.10.1"
coverage = "^5.0.3"
coverage-badge = "^1.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
